ProductDetailsGallery: Displays a Product Image Gallery
The "ProductDetailsGallery" component showcases a gallery of images for a product. It accepts a "gallery" prop, which is an array of image objects containing "large", "medium", "small", and alt properties. The component maps over the gallery array to render each image, using the <picture> element to provide responsive images based on screen size.