ConfirmationModal: New Component for Displaying Order Confirmation
This component displays the confirmation modal after an order is placed. It manages the visibility of the modal and allows users to toggle between viewing one or all items from their order. The component integrates "ConfirmationModalText", "ConfirmationModalList", and "ConfirmationModalTotal" to provide a complete overview of the order. It also includes a button for closing the modal and returning to the homepage.